Atualização final para entrega do desafio.
incrementado testes unitarios e testes automatizados

+++ back-end/src/main/java/com/se/calc/models/Item.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 
 
 @Data
@@ -13,6 +14,6 @@ public class Item {
 
     private Long id;
     private String nomeItem;
-    private Double valorItem;
+    private BigDecimal valorItem;
 
 }
